+++ backend/src/classes/Migrations/Version20190503193118.php
@@ -1,0 +1,41 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neucore\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20190503193118 extends AbstractMigration
+{
+    public function getDescription() : string
+    {
+        return '';
+    }
+
+    /**
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    public function up(Schema $schema) : void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('CREATE UNIQUE INDEX player_group_idx ON group_applications (player_id, group_id)');
+    }
+
+    /**
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    public function down(Schema $schema) : void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('DROP INDEX player_group_idx ON group_applications');
+    }
+}