+++ frontend/src/pages/AppAdmin.vue
@@ -44,7 +44,7 @@
             </div>
         </div>
         <div class="col-lg-8">
-            <ul class="nav nav-tabs">
+            <ul class="nav nav-pills nav-fill">
                 <li class="nav-item">
                     <a class="nav-link"
                        :class="{ 'active': contentType === 'managers' }"